+++ DomainModel/Enitities/Person.cs
@@ -2,18 +2,21 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.ComponentModel.DataAnnotations;
 
-namespace TestShopAspnet.Models
+
+namespace DomainModel.Enitities
 {
     public class Person
     {
+        [Key]
         public int Id { get; set; }
         public string Name { get; set; }
         public string Surname { get; set; }
         public int Age { get; set; }
         public string Position { get; set; }
 
-        private static int Increment = 1;
 
         public Person(string name, string surname, int age, string pos = "Нет информации...")
         {
@@ -21,8 +24,6 @@ namespace TestShopAspnet.Models
             Surname = surname;
             Age = age;
             Position = pos;
-            Id = Person.Increment;
-            Person.Increment++;
         }
     }
 }
