+++ TestShopAspnet/Controllers/MainController.cs
@@ -4,7 +4,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using TestShopAspnet.Models;
 using TestShopAspnet.Services.Interfaces;
 using DomainModel.Filters;
 using TestShopAspnet.ViewModels;
