+++ TestShopAspnet/Controllers/PersonsController.cs
@@ -1,9 +1,9 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using DomainModel.Enitities;
+using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using TestShopAspnet.Models;
 using TestShopAspnet.Services.Interfaces;
 using TestShopAspnet.ViewModels;
 
@@ -44,14 +44,8 @@ namespace TestShopAspnet.Controllers
             Person checkPers = _personsService.Get(id);
             if (checkPers == null)            
                 return new NotFoundResult();
-            PersonViewModel persvm = new PersonViewModel()
-            {
-                Id = checkPers.Id,
-                Name = checkPers.Name,
-                Surname = checkPers.Surname,
-                Age = checkPers.Age,
-                Position = checkPers.Position
-            };
+            PersonViewModel persvm = new PersonViewModel(checkPers);
+
             return View(persvm);
         }
 
@@ -82,14 +76,8 @@ namespace TestShopAspnet.Controllers
             if (checkPers is null)
                 return NotFound();
 
-            PersonViewModel persvm = new PersonViewModel()
-            {
-                Id = checkPers.Id,
-                Name = checkPers.Name,
-                Surname = checkPers.Surname,
-                Age = checkPers.Age,
-                Position = checkPers.Position
-            };
+            PersonViewModel persvm = new PersonViewModel(checkPers);
+            
             return View(persvm);
         }
