+++ TestShopAspnet/Services/Interfaces/IPersonsData.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using TestShopAspnet.Models;
+using DomainModel.Enitities;
 
 namespace TestShopAspnet.Services.Interfaces
 {
