+++ TestShopAspnet/Views/Persons/Info.cshtml
@@ -1,4 +1,4 @@
-﻿@model TestShopAspnet.Models.Person
+﻿@model DomainModel.Enitities.Person
 
 @{
     ViewData["Title"] = $"Сотрудник {Model.Surname} {Model.Name}";
