+++ TestShopAspnet/Views/_ViewImports.cshtml
@@ -1,5 +1,5 @@
 ﻿@using TestShopAspnet;
-@using TestShopAspnet.Models;
+@using DomainModel.Enitities;
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, TestShopAspnet
